requirements.txt
numpy
pandas>=2.0.0
pyyaml
tqdm
pyarrow
